docs

+++ src/com/brett/tools/SettingsLoader.java
@@ -61,7 +61,7 @@ public class SettingsLoader {
 				if (name[0].equals("sensitivity"))
 					SENSITIVITY = Double.parseDouble(name[1]);
 				if (name[0].equals("samples"))
-					SAMPLES = Integer.parseInt(name[1]);
+					SAMPLES = (int) Float.parseFloat(name[1]);
 				try {
 				if (name[0].equals("username"))
 					MainMenu.username = name[1].trim();
